requirements.txt
together>=1.3.0
python-dotenv>=1.0.0
tqdm>=4.66.0
datasets>=2.19.0
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0
numpy>=1.24.0
